moves the HandleDatasetAPIErr func to common errors file and renders the returned error message
