+++ api/common/dataset.go
@@ -2,8 +2,6 @@ package common
 
 import (
 	"context"
-	"net/http"
-	"strings"
 	"time"
 
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
@@ -80,29 +78,4 @@ func MapResults(results []*models.DatasetUpdate) []*models.Dataset {
 		items = append(items, item.Current)
 	}
 	return items
-}
-
-func HandleDatasetAPIErr(ctx context.Context, err error, w http.ResponseWriter, data log.Data) {
-	if data == nil {
-		data = log.Data{}
-	}
-
-	var status int
-	switch {
-	case datasetsForbidden[err]:
-		status = http.StatusForbidden
-	case datasetsNoContent[err]:
-		status = http.StatusNoContent
-	case datasetsBadRequest[err], strings.HasPrefix(err.Error(), "invalid fields:"):
-		status = http.StatusBadRequest
-	case resourcesNotFound[err]:
-		status = http.StatusNotFound
-	default:
-		err = errs.ErrInternalServer
-		status = http.StatusInternalServerError
-	}
-
-	data["responseStatus"] = status
-	log.Error(ctx, "request unsuccessful", err, data)
-	http.Error(w, err.Error(), status)
 }
